server/pyproject.toml
[project]
name = "Le-Coffre"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "bcrypt>=4.3.0",
  "fastapi[standard]==0.116.1",
  "passlib[bcrypt]>=1.7.4",
  "PyCryptodome==3.23.0",
  "pyjwt>=2.10.1",
  "sqlmodel==0.0.24",
]

[dependency-groups]
dev = [
  "pytest==7.4.4",
  "pytest-asyncio>=0.23.8",
  "pytest-watcher==0.4.3",
]
sqlite = ["sqlite-utils==3.38"]
postgres = ["psycopg2==2.9.10", "psycopg[binary]==3.2.9"]

[tool.pytest.ini_options]
pythonpath = "src"
